requirements.txt
streamlit
pandas
plotly
requests
websocket-client
pyjwt
sqlalchemy
streamlit_authenticator
PyYAML
passlib
yfinance